requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.1
astroid==2.5.1
colorama==0.4.4
defusedxml==0.7.1
diff-match-patch==20200713
dj-static==0.0.6
Django==3.0
django-import-export==2.5.0
django-suit @ https://github.com/darklow/django-suit/tarball/v2
djangorestframework==3.12.2
et-xmlfile==1.0.1
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
isort==5.7.0
lazy-object-proxy==1.5.2
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.1.0
psycopg2==2.8.6
pylint==2.7.2
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.1
PyYAML==5.4.1
reportlab==3.6.1
six==1.16.0
sqlparse==0.4.1
static3==0.7.0
tablib==3.0.0
toml==0.10.2
typed-ast==1.4.2
webencodings==0.5.1
wrapt==1.12.1
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
